provider/tutorial/requirements.txt
azureml-core
azureml-mlflow
scikit-learn
feast-azure-provider

